Separate OpenGL status and legacy GL renderer status
Before this, various parts of the source code referenced VID_MODEGL
to check either if OpenGL was active, or if the legacy OpenGL renderer
was active, because when the code was written there was only 1 OpenGL
renderer. Now that there's 2 OpenGL renderers, I had to do this to
separate them.