Fix transparent palette index
I realized the actual culprit of this is that patches are drawn without the colormap,
they just go directly into the palette, or map with a translation table into the palette.

I'm gonna have to treat the colormap like the translation table, as in there's a variation without
a map at all.