Migrate to very barebones libavformat implementation
I was gonna finish my MKV muxer, but I had to write codec private options,
and that got into the nitty gritty because I had to write an
AVCDecoderConfigurationRecord, which was specified in MPEG-4 Part 15.

The issue is that I couldn't figure out if I'd have to pay for some sort
of patent license in order to do that, I was also considering adding lavf
anyway since I didn't wanna force users to mux to mkv, in the end I decided
it'd be best to stop there.

I don't regret writing it, it was a lot of fun (albeit really hard at times)
and I learned a lot about video containers (and codecs) as a whole.