Refactor codec initialization and improve video codec selection
Before this, only one codec was used in the initialization process, and if it failed, you weren't able to viddump to that format.
Now, the program tries preferrable codecs, and also tries the codec's default codecs, and now, only if all that fails, does it error out.
I still have to refactor the audio codec, but it's roughly the same process.